Polish configuration properties
This commit adds documentation to each configuration key and makes
default values explicit so that the generated meta-data is complete.

The `remoteURL` property has been renamed to `brokerUrl` for consistency
with other messaging-based auto-configuration. Also, not suffixing it
with `URL` prevents the canonical property name to be `remote-u-r-l`.

Finally, blacklist and whitelist are lists so they are now defined as
such in the configuration. That way it can be configured hierarchically
with yml.